chore(application): add authentication controller file
- reduce the length of the user controller file by creating an authentication controller file to handle signin and signup processes

[finishes #153779688]
