+++ server/controller/index.js
@@ -4,6 +4,7 @@ import borrowlog from './Borrowlog';
 import bookcategory from './Category';
 import usertype from './UserType';
 import accounttype from './AccountType';
+import authentication from './Authentication';
 
 /**
  * Export all the controllers in this folder
@@ -14,5 +15,6 @@ export default {
   borrowlog,
   usertype,
   accounttype,
-  bookcategory
+  bookcategory,
+  authentication
 };
